fix: created PayLaterAvailabilityChecker, created AvailabilityContextBuilder, updated tests, implemented checkers and builders
